+++ NEWS.md
@@ -5,6 +5,8 @@
 * Fixed bugs in `as_tsibble.ts()` for monthly series that not start at January. (#89)
 * `guess_frequency.yearweek()` returns 52.18 for more accurate weekly representation, instead of 52.
 * `n()` now can be called in `slice.tbl_ts()`. (#95)
+* Compute interval of "day" for `difftime`.
+* Fixed interval of "day" for `hms`.
 
 ## Misc